+++ app/src/androidTest/java/ch/epfl/sweng/zuluzulu/AssociationFragmentAsAuthenticatedTest.java
@@ -1,0 +1,48 @@
+package ch.epfl.sweng.zuluzulu;
+
+import android.support.test.espresso.action.ViewActions;
+import android.support.test.rule.ActivityTestRule;
+import android.support.test.runner.AndroidJUnit4;
+
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import java.util.concurrent.TimeUnit;
+
+import ch.epfl.sweng.zuluzulu.Fragments.AssociationFragment;
+import ch.epfl.sweng.zuluzulu.Structure.AuthenticatedUser;
+import ch.epfl.sweng.zuluzulu.Structure.User;
+
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+
+@RunWith(AndroidJUnit4.class)
+public class AssociationFragmentAsAuthenticatedTest {
+
+    @Rule
+    public final ActivityTestRule<MainActivity> mActivityRule =
+            new ActivityTestRule<>(MainActivity.class);
+
+    @Before
+    public void initAuthenticatedTest(){
+        User user = Utility.createTestUser();
+        AssociationFragment fragment = AssociationFragment.newInstance(user);
+        mActivityRule.getActivity().openFragment(fragment);
+
+        try {
+            TimeUnit.SECONDS.sleep(2);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Test
+    public void authenticatedClickOnFavorites() {
+        initAuthenticatedTest();
+        onView(withId(R.id.association_fragment_fav_button)).perform(ViewActions.click());
+    }
+
+
+}
