+++ app/src/main/java/ch/epfl/sweng/zuluzulu/MainActivity.java
@@ -277,4 +277,8 @@ public class MainActivity extends AppCompatActivity implements OnFragmentInterac
     public User getUser() {
         return user;
     }
+
+    public void setUser(User user){
+        this.user = user;
+    }
 }
